requirements.txt
matplotlib
numpy
pandas
pillow
timm
torch
torchvision
tqdm